It changes the SQL a lot, but I kept all existing checks

- don't query files torrents are not inserted
- use basic types as much as possible (no strftime, hex)
- run sqlite in batch instead of window

Here are my results, ran on AMD Ryzen Pro 4750G + 16GB RAM on SSD, with a 29 millions magnetico database. Arguments list: `--source-name magnetico --add-files --add-files-limit 200 --insert-content`
- On an empty postgresql target: from 260 it/s to 980 it/s
- On an already filled postgreql target (so a lot of conflicts): from 260 it/s (unchanged) to 6000 it/s

So it's a huge win for pre-filled postgresql target, since this will never query.

**Please double check that I didn't miss something.**
I didn't added #10 in this, but I ran my tests with it.