@Glandos 2025-02-26
I've added support for pgcopy.

The code is somewhat more complicated, but performance are nearly doubled.
The only missing bit is insertion in `torrent_contents` because it uses a `tsvector` which is not supported by pgcopy. Not yet: https://github.com/altaurog/pgcopy/pull/47