@Glandos 2025-02-26
Using pgcopy for file insertion removed the detection of duplicated file that are all empty. I don't know if I can restore that easily, since it's done in a generator.