@Glandos 2025-03-01
|Target DB status | `main` output | `pg_perf` output |
| --- | --- | --- |
| empty | 2282/29000001 [00:30<107:59:03, 74.59it/s] | 58000/29000001 [00:30<4:15:12, 1890.13it/s]
| full | 9811/29000001 [00:30<25:05:19, 320.97it/s] | 226000/29000001 [00:30<1:05:05, 7367.77it/s]

"empty" is a database without any entries. "full" is the database with all entries existing. This is useful for re-importing entries with only a few changes, e.g. merging divergent data.

The benchmark was run with:
```
find ./bitmagnet/base/ -type f -exec dd if={} iflag=nocache count=0 status=none \; && systemctl restart postgresql@17-bitmagnet
dd if=./database.sqlite3 iflag=nocache count=0; uv run --with 'pgcopy @ https://github.com/Glandos/pgcopy/archive/refs/heads/tsvector.zip' --with 'tqdm' magnetico2database.py --dbname bitmagnet_full --user bitmagnet --password 'XXX' --host localhost --port 5433 --source-name magnetico --add-files --add-files-limit 200 --insert-content ./database.sqlite3
```
- `dd`  is useful to evict file cache.
- `pgcopy` is my own branch, waiting for merging
- I replaced `bitmagnet_full` with `bitmagnet_empty` as dbname for the other test.